refactor: split Ct catalog widgets and map-state forwarders (#3878)
Extract CtCompassRose painter, CtDropdown trigger/picker parts, diplomacy
badge chrome into section/relation parts, and GameMapAreaStateLogic API
forwarders to keep host files lean without changing public behavior.
