+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel.dart
@@ -29,7 +29,8 @@ export 'diplomacy_panel_rows.dart';
 export 'relative_power_line.dart';
 
 part 'diplomacy_panel_body.dart';
-part 'diplomacy_panel_chrome_badges.dart';
+part 'diplomacy_panel_chrome_section_header.dart';
+part 'diplomacy_panel_chrome_relation_badges.dart';
 part 'diplomacy_panel_chrome_standing.dart';
 part 'diplomacy_panel_constants.dart';
 part 'diplomacy_panel_filter.dart';
